+++ src/module/item/background.ts
@@ -1,5 +1,5 @@
 import { PF2ECharacter } from '../actor/character';
-import { ABCFeatureEntryData, BackgroundData, FeatData } from './dataDefinitions';
+import { ABCFeatureEntryData, BackgroundData, FeatData } from './data-definitions';
 import { PF2EItem } from './item';
 import { PF2EFeat } from './others';
 
@@ -8,14 +8,17 @@ export class PF2EBackground extends PF2EItem {
 
     static async getBackgroundItemData(entry: ABCFeatureEntryData): Promise<FeatData> {
         if (entry.pack) {
-            const pack = game.packs.get<PF2EFeat>(entry.pack);
-            return pack.getEntry(entry.id);
+            const pack = game.packs.get<Compendium<PF2EFeat>>(entry.pack);
+            const entity = await pack.getEntity(entry.id);
+            return entity instanceof PF2EFeat
+                ? entity._data
+                : Promise.reject(new Error('Invalid item type referenced in ABCFeatureEntryData'));
         } else {
             const feat = game.items.get(entry.id);
             if (feat?.data.type !== 'feat') {
-                throw Error('Invalid item type referenced in ABCFeatureEntryData');
+                return Promise.reject(new Error('Invalid item type referenced in ABCFeatureEntryData'));
             }
-            return duplicate(feat?.data);
+            return duplicate(feat.data);
         }
     }