+++ src/module/item/data-definitions.ts
@@ -1,5 +1,5 @@
-import { AbilityString, Proficency } from '../actor/actorDataDefinitions';
-import { PF2RuleElementData } from '../rules/rulesDataDefinitions';
+import { AbilityString, Proficency } from '@actor/actor-data-definitions';
+import { PF2RuleElementData } from '../rules/rules-data-definitions';
 
 export type Size = 'tiny' | 'sm' | 'med' | 'lg' | 'huge' | 'grg';
 
@@ -551,6 +551,9 @@ export interface SpellDetailsData extends ItemDescriptionData {
     };
     heightenedLevel: {
         value: number;
+    };
+    hasCounteractCheck: {
+        value: boolean;
     };
     isSave?: boolean;
     damageLabel?: string;