+++ src/module/item/effect.ts
@@ -1,5 +1,5 @@
 import { PF2EItem } from './item';
-import { EffectData } from './dataDefinitions';
+import { EffectData } from './data-definitions';
 
 export class PF2EEffect extends PF2EItem {
     data!: EffectData;