+++ src/scripts/chat/crit-fumble-cards.ts
@@ -4,9 +4,9 @@ class PF2eCritFumbleCards {
     static diceSoNice: boolean;
 
     static async init() {
-        const rollableTables = game.packs.get('pf2e.rollable-tables');
-        this.critTable = new RollTable(await rollableTables.getEntry('FTEpsIWWVrDj0jNG'));
-        this.fumbleTable = new RollTable(await rollableTables.getEntry('WzMGWMIrrPvSp75D'));
+        const rollableTables = game.packs.get<Compendium<RollTable>>('pf2e.rollable-tables');
+        this.critTable = (await rollableTables.getEntity('FTEpsIWWVrDj0jNG'))!;
+        this.fumbleTable = (await rollableTables.getEntity('WzMGWMIrrPvSp75D'))!;
 
         if (game.settings.get('pf2e', 'drawCritFumble')) {
             // Support diceSoNice module